move mock model to common collection
